maybe need f16 specified?

+++ misc/convobot.py
@@ -940,7 +940,7 @@ class Bot:
 					if rem > 8 * 1024:
 						from transformers import BitsAndBytesConfig
 						quantization_config = BitsAndBytesConfig(llm_int8_enable_fp32_cpu_offload=True)
-						model = backup_model(AutoModelForCausalLM.from_pretrained, m, device_map=dev_map, load_in_8bit=True, quantization_config=quantization_config)
+						model = backup_model(AutoModelForCausalLM.from_pretrained, m, device_map=dev_map, torch_dtype=torch.float16, load_in_8bit=True, quantization_config=quantization_config)
 					else:
 						model = backup_model(AutoModelForCausalLM.from_pretrained, m, device_map=dev_map, load_in_8bit=True)
 				print(dev_map)
